feat: the second derivative of an analytic map is symmetric (#19770)
Unify the facts that the second derivative of a `C^2` function over R or C, or an analytic function over an arbitrary field, is symmetric. This is expressed as an assumption that `f` is `C^n` for some `n ≥ minSmoothness 𝕜 2`, where `minSmoothness 𝕜 i` is by definition `i` if the field is R or C, and `ω` otherwise.