feat(RingTheory/Ideal/Over): define the set of all prime ideals that lie over an ideal (#19139)
Define the set of all prime ideals that lie over an ideal. It is mainly for #19141.

This is adapted from the proof in [the case of number fields](https://github.com/jjdishere/neukirch/blob/8fdb75f7ebab8a20eea902058956d7dece648959/AlgebraicNumberTheory/AlgebraicIntegersPart2/HilbertRamificationTheory.lean#L384).

Thanks to @erdOne for reminding me that something similar was done in `flt-regular`. I referred to it to make my code a lot simpler.



Co-authored-by: Hu Yongle <mbkybky@qq.com>
Co-authored-by: Yongle Hu <140475041+mbkybky@users.noreply.github.com>
Co-authored-by: Yongle Hu <2065545849@qq.com>