refactor: add an `ofNat()` elaborator
`ofNat(n)` is a shorthand for `no_index (OfNat.ofNat n)`.

Its purpose is:
* To make it easier to remember to write the `no_index`
* To add a place to put a hoverable docstrings explaining the complexities
* To make it easier to remove the `no_index`es in future if `DiscrTree`s stops needing them around `ofNat`.

This is not exhaustive, but is a step in the right direction.
For now, I only target (some of) the declarations with a `See note [no_index around OfNat.ofNat]`.

Some theorem statements have been corrected in passing.