feat(RingTheory/HahnSeries/Summable): define summable smulFamily (#20189)
This PR defines a summable family of Hahn series given another summable family and a parametrized family of scalars.  The basic use case is that we already have `powers`, which gives a summable family of non-negative powers of a positive-order Hahn series, and we want to make a family whose powers are multiplied by the coefficients of a formal power series.  This will (eventually) give us an `R`-algebra map from `PowerSeries R` to `HahnSeries Γ R`.