chore(CategoryTheory/Shift): fix NatTrans.CommShift (#20364)
This PR improves some proofs about `NatTrans.CommShift` (the commutation of a natural transformation to shifts). The field `comm'` of `NatTrans.CommShift` had an unnecessary `'` in it, it is renamed `shift_comm`, and API lemmas are moved from `NatTrans.CommShift` to the `NatTrans` namespace (where `comm` becomes `shift_comm` and `comm_app` becomes `shift_app_comm`).