chore(Ideal/Basic): dependent generalization of `Ideal.pi` (#20535)
Also golf some proofs.

TODO: similar to `Ideal.idealProdEquiv`, we may construct OrderEmbedding from `Πᵢ Ideal Rᵢ` to `Ideal (Πᵢ Rᵢ)`, which is an OrderIso if the indexing type is finite. This can then be used to show e.g. that a finite product of Noetherian rings is Noetherian without using an inductive argument.