ci: basic check on fields in 100.yaml and 1000.yaml (#20078)
This PR makes `scripts/yaml_check.py` verify that `100.yaml` and `1000.yaml` only contain fields the webpage expects them to. Otherwise, changes in mathlib breaking the website won't be noticed until they are merged; this has happened several times in the last days. In particular, replace two occurrences of `note` in `1000.yaml` by `comment`.
To this end, this PR copies the dataclass constructors which are used in the website script to `scripts/yaml_check.py` (along with a comment asking to keep them in sync).

While at it, this PR documents the fields in `100.yaml` as well, and moves the documentation for `1000.yaml` to the copied dataclasses in `yaml_check.py`.

In addition, validate that at most one of the fields `decl` and `decls` are passed: the webpage also does this; this avoids late surprises.

Co-authored-by: grunweg <rothgang@math.uni-bonn.de>



Co-authored-by: Michael Rothgang <rothgang@math.uni-bonn.de>