refactor(lint-style): make sure to read the nolints file outside of t… (#19384)
…he linter implementation

This way, we categorically prevent cache invalidation bugs caused by lake replaying unchanged inputs, and thus not reaction to changes in the linter exceptions. For *syntax linters*, such a bug actually occurred in mathlib. (With the current implementation of text-based linters, this issue won't occur; this is merely a defensive change.)

Suggested in #19275.



Co-authored-by: Jon Eugster <eugster.jon@gmail.com>