chore: remove --loc flag from unused_in_pole.sh (#20729)
When developing `unused_in_pole.sh`, the Speed Center was broken, so I used the `--loc` flag to compute the longest pole according to lines of code rather than instruction counts. Let's remove it again, as I think instructions are more useful, and that's what we usually do when looking at the longest pole.