feat: isometries from induced metrics (#20731)
If `f : α → β`, where `β` is a metric space, and the metric on `α` is induced by pullback along `f`, then `f` is an isometry. I add this trivial lemma for `PseudoEmetricSpace`, `PseudoMetricSpace`, `EMetricSpace`, and `MetricSpace`, as suggested by Junyan Xu in [this thread](https://leanprover.zulipchat.com/#narrow/channel/217875-Is-there-code-for-X.3F/topic/dist.20of.20induced.20metric/near/493459925).