refactor: rename `UniqueContinuousFunctionalCalculus` to `ContinuousMap.UniqueHom` (#20643)
Given that `UniqueContinuousFunctionalCalculus` no longer mentions `spectrum`, it really has very little to tie it to the ContinuousFunctionalCalculus. In some sense, it's just a property of `C(s, R)`, but because we want to allow `R := ℝ≥0`, the algebra `A` also matters. Because of this, we put the class within the `ContinuousMap` namespace.

We also rename `UniqueNonUnitalContinuousFunctionalCalculus` to `ContinuousMapZero.UniqueHom`.