feat: remove `UniqueDiffWithinAt` assumptions for one-dimensional derivative (#20816)
A point in a subset of a field is either a point of unique differentials, or isolated. In both cases, there is a definite value for the derivative of a map (its unique derivative in the former case, zero in the latter case). This means there is no choice involved, contrary to what happens in higher dimension. This remark means we can remove many `UniqueDiffWithinAt` assumptions in one-dimensional settings.