feat: start building an API for vector fields on manifolds (#20639)
Notably, define the pullback of vector fields, and show that the pullback of a differentiable vector field under a C^2 map is differentiable.

More API to come in further PRs. Once all the API is there, it will be easy to define the Lie algebra of a Lie group using left-invariant vector fields, in #18396.