feat: `notation3`-defined pretty printers now make tokens that are clickable in documentation
Modifies the `notation3` delaborator generator so that when `pp.tagAppFns` is true, and when the expression being delaborated is the application of a constant, all the tokens of the notation are annotated with this constant. The consequence is that in docgen documentation, the tokens will be links to that constant.

In complicated notations this heuristic may be incorrect. For example, it's possible to use `notation3` to define notations that have different head constants in different expansions. There are no such cases of this in mathlib.

This heuristic might also give unexpected results. For example, in
```lean
notation3 "Nat≥0" => { x : Nat // 0 ≤ x }
```
the documentation will have appearances of `Nat≥0` link to `Subtype`.

[Zulip discussion](https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/Filter.2EEventually.20.28.60.E2.88.80.E1.B6.A0.60.29.20notation.20is.20not.20clickable.20in.20docs/near/494590979)