chore(Data/List/Basic): split off Map2 lemmas (#20832)
This PR splits off a number of lemmas related maps that zip pairs of lists from `Data.List.Basic` to a new file, including:

* `map₂Left'`
* `map₂Right'`
* `zipWith`
* `zipLeft'`
* `zipRight'`

Doing so reduces the size of this long file by ~250 lines. Also note that the new file doesn't need to import `Basic`, so this should parallelize the compilation too.