feat(CategoryTheory): the equivalence of categories induced by a bijection (#20881)
When `f : T → D` is a map and `D` is a category, then `InducedCategory D f` is a category with objects `T` with a fully faithful functor to `D`. In this PR, we show that in the case of a bijection `T ≃ D`, the induced category is equivalent to `D`.

The code would be nicer (see #19945) if the type of morphisms in the induced category was defined as a 1-field structure. This would be a very welcomed refactor.