feat(Order): relate chains to antichains (#20756)
A set which is a chain and an antichain is subsingleton. As a result, the intersection of a chain and antichain is subsingleton.

Useful from the disproof of the Aharoni–Korman conjecture, https://github.com/leanprover-community/mathlib4/pull/20082.

I believe the large import here is acceptable, as `IsChain` and `IsAntichain` need to be related to each other somewhere, but there is not much to say about both (as this PR shows). As such, making a third file specifically for these seems unnecessary, and putting these in `Order.Chain` results in a larger import change.