feat: interval integrability of periodic functions (#20938)
Add a theorem showing that a periodic function is interval integrable over every interval if it is interval integrable
over one period. Add docstrings with section headings to improve readability. The theorem is used in [Project VD](https://github.com/kebekus/ProjectVD), which aims to formalize Value Distribution Theory for meromorphic functions on the complex plane.