refactor(SetTheory/Game/Impartial): define in terms of Identical over Equiv
(In an effort to split out #7162) we define impartial using the identical relation instead of the equivalence relation, which, as proved by the Sprague-Grundy theorem in Nim.lean, is equivalent to our original definition.

Since this splits #7162, we rewrite impartial_congr in terms of Identical instead of Relabelling, as the eventual goal of #7162 is to deprecate Relabelling.

Co-authored-by: Yuyang Zhao 赵雨扬 <zhao.yu-yang@foxmail.com>