chore(AlgebraicTopology): split SimplicialSet.Basic (#21025)
This PR only splits the file `AlgebraicTopology.SimplicialSet.Basic` by creating three new files `StdSimplex`, `Boundary`, `Horn`. This prepares for a refactor and future developments of API for these simplicial sets. For example, the boundary and the horn should be redefined (keeping the defeq) as subcomplexes of the standard simplex using the `Subpresheaf` API, see #20840.