feat: do not allocate heap cells for Erased (#21068)
The memory representation of `Erased` used to be a heap cell, because `Erased` is defined in terms of `PSigma`.  If we define `Erased` as a subtype of a predicate instead, then the compiler will erase values of type `Erased`.

Unfortunately, this does not work for ghost fields in structures (yet).  The structure will still contain a field for the ghost value, but at least it will be the scalar zero now.  But I suspect it would be fairly small diff to fix this in core.

```lean
structure Demo where
  a : Erased Nat
  b : Erased Nat

set_option trace.compiler.ir.result true in
def foo : Demo where a := .mk 42; b := .mk 8
```

Before:
```
[result]
def foo._closed_1 : obj :=
  let x_1 : obj := ctor_0[PSigma.mk] ◾ ◾;
  inc x_1;
  let x_2 : obj := ctor_0[Demo.mk] x_1 x_1;
  ret x_2
def foo : obj :=
  let x_1 : obj := foo._closed_1;
  ret x_1
```

After:
```
[result]
def foo : obj :=
  let x_1 : obj := ctor_0[Demo.mk] ◾ ◾;
  ret x_1
```

@digama0 This is what I was suggesting earlier this week.