chore(1000.yaml): less ambiguous titles
Re-generate the theorem titles from upstream, using a slightly better algorithm.
This fixes a TODO (also remarked by Michael Stoll), and makes editing the file
easier since it is clearer which result is meant.

This catches one declaration name which did not refer to the right theorem
(it referred to a different version of Tonelli's theorem).

While at it, remove one further outdated TODO, and mention the PNT+
repository in a comment.